home/appCollector/requirements.txt
lywsd03mmc
paho-mqtt
